+++ src/main/java/de/mari_023/ae2wtlib/mixin/TrinketUniversalTerminalMixin.java
@@ -13,6 +13,6 @@ import dev.emi.trinkets.api.Trinket;
 @Mixin(value = ItemWUT.class, remap = false)
 public abstract class TrinketUniversalTerminalMixin extends ItemWT implements Trinket {
     public void tick(ItemStack stack, SlotReference slot, LivingEntity entity) {
-        inventoryTick(stack, entity.getLevel(), entity, 0, false);
+        inventoryTick(stack, entity.level(), entity, 0, false);
     }
 }