+++ src/main/java/de/mari_023/ae2wtlib/terminal/IUniversalWirelessTerminalItem.java
@@ -1,21 +1,16 @@
 package de.mari_023.ae2wtlib.terminal;
 
-import java.util.OptionalLong;
-
 import org.jetbrains.annotations.Nullable;
 
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
 
 import de.mari_023.ae2wtlib.AE2wtlib;
 import de.mari_023.ae2wtlib.wut.WUTHandler;
 
-import appeng.api.features.Locatables;
 import appeng.api.implementations.menuobjects.ItemMenuHost;
-import appeng.api.networking.security.IActionHost;
-import appeng.core.localization.PlayerMessages;
 import appeng.menu.MenuOpener;
 import appeng.menu.locator.MenuLocator;
 
@@ -44,31 +39,6 @@ public interface IUniversalWirelessTerminalItem {
             return false;
         }
 
-        Level level = player.getCommandSenderWorld();
-        if (level.isClientSide()) {
-            return false;
-        }
-
-        OptionalLong key = getGridKey(item);
-        if (key.isEmpty()) {
-            player.sendSystemMessage(PlayerMessages.DeviceNotLinked.text());
-            return false;
-        }
-
-        IActionHost securityStation = Locatables.securityStations().get(level, key.getAsLong());
-        if (securityStation == null) {
-            player.sendSystemMessage(PlayerMessages.StationCanNotBeLocated.text());
-            return false;
-        }
-
-        if (!hasPower(player, 0.5, item)) {
-            player.sendSystemMessage(PlayerMessages.DeviceNotPowered.text());
-            return false;
-        }
-        return true;
+        return player.level() instanceof ServerLevel;
     }
-
-    OptionalLong getGridKey(ItemStack item);
-
-    boolean hasPower(Player player, double amt, ItemStack is);
 }