+++ src/main/java/de/mari_023/ae2wtlib/wct/magnet_card/MagnetHandler.java
@@ -78,20 +78,20 @@ public class MagnetHandler {
     private static void handleMagnet(Player player, ItemStack terminal) {
         if (player.isShiftKeyDown())
             return;
-        if (!getMagnetSettings(terminal).isActive())
+        if (getMagnetSettings(terminal).isDisabled())
             return;
         handleMagnet(player);
     }
 
     private static void handleMagnet(Player player) {
         CraftingTerminalHandler ctHandler = CraftingTerminalHandler.getCraftingTerminalHandler(player);
-        if (!getMagnetSettings(ctHandler.getCraftingTerminal()).isActive())
+        if (getMagnetSettings(ctHandler.getCraftingTerminal()).isDisabled())
             return;
         MagnetHost magnetHost = ctHandler.getMagnetHost();
         if (magnetHost == null)
             return;
 
-        List<ItemEntity> entityItems = player.getLevel().getEntitiesOfClass(ItemEntity.class,
+        List<ItemEntity> entityItems = player.level().getEntitiesOfClass(ItemEntity.class,
                 player.getBoundingBox().inflate(AE2wtlibConfig.INSTANCE.magnetCardRange()),
                 EntitySelector.ENTITY_STILL_ALIVE);