+++ src/main/java/de/mari_023/ae2wtlib/wet/WETMenuHost.java
@@ -52,7 +52,7 @@ public class WETMenuHost extends WTMenuHost
 
     @Override
     public Level getLevel() {
-        return getPlayer().getLevel();
+        return getPlayer().level();
     }
 
     @Override