+++ src/main/java/de/mari_023/ae2wtlib/wut/CycleTerminalButton.java
@@ -4,10 +4,10 @@ import java.util.Collections;
 import java.util.List;
 
 import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.PoseStack;
 
 import org.jetbrains.annotations.NotNull;
 
+import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.renderer.Rect2i;
 import net.minecraft.network.chat.Component;
@@ -46,28 +46,25 @@ public class CycleTerminalButton extends Button implements ITooltip {
             "textures/item/wireless_universal_terminal.png");
 
     @Override
-    public void renderWidget(PoseStack matrices, final int mouseX, final int mouseY, float partial) {
+    public void renderWidget(GuiGraphics guiGraphics, int mouseX, int mouseY, float partial) {
         if (!visible)
             return;
-        matrices.pushPose();
-        RenderSystem.setShaderTexture(0, TEXTURE_STATES);
+        guiGraphics.pose().pushPose();
         RenderSystem.disableDepthTest();
         RenderSystem.enableBlend();
         if (isFocused()) {
-            fill(matrices, getX() - 1, getY() - 1, getX() + width + 1, getY() + height + 1, 0xFFFFFFFF);
+            guiGraphics.fill(getX() - 1, getY() - 1, getX() + width + 1, getY() + height + 1, 0xFFFFFFFF);
         }
 
-        blit(matrices, getX(), getY(), 240, 240, 16, 16);
-
-        RenderSystem.setShaderTexture(0, nextTerminal);
+        guiGraphics.blit(TEXTURE_STATES, getX(), getY(), 240, 240, 16, 16);
 
         if (active)
             RenderSystem.setShaderColor(1.0f, 1.0f, 1.0f, 1.0f);
         else
             RenderSystem.setShaderColor(0.5f, 0.5f, 0.5f, 1.0f);
 
-        blit(matrices, getX() + 1, getY() + 1, 14, 14, 0, 0, 512, 512, 512, 512);
+        guiGraphics.blit(nextTerminal, getX() + 1, getY() + 1, 14, 14, 0, 0, 512, 512, 512, 512);
 
-        matrices.popPose();
+        guiGraphics.pose().popPose();
     }
 }